feat: add error implementation and implement an better SoC
