+++ apps/web/src/adapters/views/templates/file-error.template.ts
@@ -1,0 +1,8 @@
+export function fileErrorTemplate(tipMessage: string) {
+  const tip = document.getElementById("tip")!;
+  const dropZone = document.getElementById("drop-zone")!;
+  tip.classList.add("error");
+  tip.innerText = tipMessage;
+  dropZone.appendChild(tip);
+  dropZone.classList.add("error");
+}
