+++ apps/web/src/adapters/views/validators/file-not-exists.validators.ts
@@ -1,0 +1,10 @@
+import { fileErrorTemplate } from "../templates/file-error.template";
+
+export function fileNotExistValidator(file: File) {
+  if (!file) {
+    fileErrorTemplate("No files selected.");
+    return false;
+  }
+
+  return true;
+}
