+++ apps/web/src/adapters/views/validators/file-not-json.validators.ts
@@ -1,0 +1,10 @@
+import { fileErrorTemplate } from "../templates/file-error.template";
+
+export function fileNotJsonValidator(file: File) {
+  if (!file.name.endsWith(".json")) {
+    fileErrorTemplate("The selected file is not an JSON.");
+    return false;
+  }
+
+  return true;
+}
