+++ apps/web/src/core/constants/default-messages.contants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_TIP_MESSAGE =
+  "Simple JSON Viewer that runs completely on-client. No data exchange";
